feat: move the request header support to the new codebase
