+++ packages/mock-addon/src/utils/faker.js
@@ -1,12 +1,13 @@
 /* eslint-disable no-restricted-globals */
 import { newMockXhr } from 'mock-xmlhttprequest';
 import { match } from 'path-to-regexp';
-import { Request, setRequestHeaders } from './request';
+import { Request } from './request';
+import { Response } from './response';
 import {
-    Response,
+    setRequestHeaders,
     getResponseHeaderMap,
     defaultResponseHeaders,
-} from './response';
+} from './headers';
 import { arrayEquals } from './array';
 import { getNormalizedUrl } from './url';
 
@@ -142,7 +143,8 @@ export class Faker {
             const realXhr = new global.realXMLHttpRequest();
             realXhr.open(method, url);
 
-            setRequestHeaders(realXhr, xhr.requestHeaders);
+            setRequestHeaders(realXhr, xhr.requestHeaders._headers);
+            realXhr.withCredentials = xhr._withCredentials;
 
             realXhr.onreadystatechange = function onReadyStateChange() {
                 if (realXhr.readyState === 4 && realXhr.status === 200) {
